+++ frontend/src/containers/NewOrEditPost/hooks/useNewPost.ts
@@ -1,0 +1,6 @@
+import { useContext } from "react";
+import { NewPostContext } from "../contexts/NewPostContext";
+
+export const useNewPost = () => {
+  return useContext(NewPostContext);
+};
